+++ src/com/company/menu/SpecialMenu.java
@@ -3,9 +3,9 @@ package com.company.menu;
 import java.time.LocalDate;
 
 public class SpecialMenu extends Menu {
-    int discount = 0;
+    double discount = 0.0;
 
-    public SpecialMenu(int id, String name, String description, LocalDate date_created, int discount){
+    public SpecialMenu(int id, String name, String description, LocalDate date_created, double discount){
         super(id, name, description, date_created);
         this.discount = discount;
     }
